+++ app/models/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+# Example Pydantic model for creating items
+class Item(BaseModel):
+    symbol: str
+    exchange: str
+    exchangeShortName: str
+    name: str
+    type: str
